requirements.txt
Flask==3.1.0
flask_cors==5.0.1
langchain_chroma==0.2.2
langchain_community==0.3.20
langchain_ollama==0.3.0
langchain_text_splitters==0.3.7
ollama==0.4.7
pymongo==4.11.3
python-dotenv==1.1.0
Werkzeug==3.1.3
